requirements.txt
django
djangorestframework
channels
django-cors-headers
drf-nested-routers
daphne